+++ src/main/java/com/gregtechceu/gtceu/api/sync_system/ClassSyncData.java
@@ -6,19 +6,19 @@ import com.gregtechceu.gtceu.api.sync_system.annotations.SaveField;
 import com.gregtechceu.gtceu.api.sync_system.annotations.SyncToClient;
 import com.gregtechceu.gtceu.api.sync_system.data_transformers.ValueTransformer;
 import com.gregtechceu.gtceu.api.sync_system.data_transformers.ValueTransformers;
-import com.gregtechceu.gtceu.api.sync_system.managed.ISyncAnnotated;
 import com.gregtechceu.gtceu.api.sync_system.managed.ISyncManaged;
 
 import it.unimi.dsi.fastutil.objects.ObjectArrayList;
 import it.unimi.dsi.fastutil.objects.ObjectOpenHashSet;
 import lombok.Getter;
-import lombok.SneakyThrows;
+import org.jetbrains.annotations.Nullable;
 
 import java.lang.invoke.MethodHandle;
 import java.lang.invoke.MethodHandles;
 import java.lang.invoke.VarHandle;
 import java.lang.reflect.*;
 import java.util.*;
+import java.util.function.Supplier;
 
 /**
  * Static data for {@link ISyncManaged} classes.
@@ -47,23 +47,21 @@ public final class ClassSyncData {
     private final Set<FieldSyncData> clientSyncFields = new ObjectOpenHashSet<>();
     @Getter
     private final Set<FieldSyncData> serverSaveFields = new ObjectOpenHashSet<>();
+    @Getter
+    private @Nullable Supplier<?> clientsideConstructor = null;
 
-    @SneakyThrows
     private ClassSyncData(Class<?> clazz) {
         var isManaged = ISyncManaged.class.isAssignableFrom(clazz);
-        var isAnnotated = ISyncAnnotated.class.isAssignableFrom(clazz);
 
-        if (!isManaged && !isAnnotated)
-            throw new IllegalArgumentException("Cannot create class sync data for non-sync class");
-        if (isManaged && isAnnotated) throw new IllegalArgumentException(
-                "Class %s cannot inherit both ISyncAnnotated and ISyncManaged".formatted(clazz.getName()));
+        if (!isManaged)
+            throw new IllegalArgumentException("Cannot create class sync data for non sync managed class.");
 
         MethodHandles.Lookup privateLookup;
         try {
             privateLookup = MethodHandles.privateLookupIn(clazz, LOOKUP);
         } catch (IllegalAccessException e) {
             GTCEu.LOGGER.error("Sync: Failed to create method handle lookup for class {}", clazz);
-            throw e;
+            throw new RuntimeException(e);
         }
 
         Map<String, List<MethodHandle>> changeListeners = new HashMap<>();
@@ -105,7 +103,7 @@ public final class ClassSyncData {
             } catch (IllegalAccessException e) {
                 GTCEu.LOGGER.error("Sync: Failed to acquire variable handle for field {} {}", field.getName(),
                         clazz.getName());
-                throw e;
+                throw new RuntimeException(e);
             }
 
             FieldSyncData syncData = new FieldSyncData(field, handle, ValueTransformers.get(field.getGenericType()),
@@ -116,7 +114,7 @@ public final class ClassSyncData {
         }
 
         Class<?> parent = clazz.getSuperclass();
-        if (ISyncManaged.class.isAssignableFrom(parent) || ISyncAnnotated.class.isAssignableFrom(parent)) {
+        if (ISyncManaged.class.isAssignableFrom(parent)) {
             ClassSyncData parentHandles = CACHE.get(parent);
             managedFields.addAll(parentHandles.managedFields);
             clientSyncFields.addAll(parentHandles.clientSyncFields);
@@ -135,5 +133,13 @@ public final class ClassSyncData {
         managedFields.stream().filter(f -> Objects.equals(f.fieldName, fieldName))
                 .findFirst()
                 .ifPresent(fieldData -> fieldData.setTransformer(transformer));
+    }
+
+    /**
+     * Allows this class to be instantiated on the client when being synced.
+     * @param ctor The constructor/supplier to be called when instantiating this class on the client.
+     */
+    public void setClientsideConstructor(@Nullable Supplier<?> ctor) {
+        clientsideConstructor = ctor;
     }
 }
