+++ src/main/java/com/gregtechceu/gtceu/api/sync_system/SyncDataHolder.java
@@ -2,6 +2,7 @@ package com.gregtechceu.gtceu.api.sync_system;
 
 import com.gregtechceu.gtceu.GTCEu;
 import com.gregtechceu.gtceu.api.sync_system.data_transformers.ValueTransformer;
+import com.gregtechceu.gtceu.api.sync_system.data_transformers.ValueTransformers;
 import com.gregtechceu.gtceu.api.sync_system.managed.ISyncManaged;
 
 import net.minecraft.nbt.CompoundTag;
@@ -57,7 +58,7 @@ public class SyncDataHolder {
         CompoundTag tag = new CompoundTag();
         for (var field : fieldsToSerialize) {
             if (shouldSerializeField(field, writeClientFields, fullSync)) {
-                Tag nbtValue = FieldSyncHandler.serializeField(holder, field, writeClientFields, fullSync);
+                Tag nbtValue = serializeField(holder, field, writeClientFields, fullSync);
                 tag.put(field.nbtSaveKey, nbtValue);
             }
         }
@@ -78,7 +79,7 @@ public class SyncDataHolder {
         for (var field : fieldsToCheck) {
 
             Tag savedValue = tag.get(field.nbtSaveKey);
-            FieldSyncHandler.deserializeField(holder, field, savedValue, readingClientFields);
+            deserializeField(holder, field, savedValue, readingClientFields);
 
             if (readingClientFields) {
                 try {
@@ -99,6 +100,80 @@ public class SyncDataHolder {
         }
     }
 
+    @SuppressWarnings("unchecked")
+    private static Tag serializeField(Object holder, FieldSyncData field,
+                                     boolean writeClientFields, boolean fullSync) {
+        Object currentValue = field.handle.get(holder);
+
+        if (currentValue == null) {
+            var nullCompound = new CompoundTag();
+            nullCompound.putBoolean("null", true);
+            return nullCompound;
+        }
+
+        if (field.transformer == null) {
+            field.setTransformer(ValueTransformers.get(field.type.getRawType()));
+            if (field.transformer == null) {
+                GTCEu.LOGGER.error("Sync: Failed to serialize field {} in class {}: Missing value transformer for {}",
+                        field.fieldName, holder.getClass().getName(), field.type);
+                return new CompoundTag();
+            }
+        }
+
+        try {
+            return ((ValueTransformer<Object>) field.transformer).serializeNBT(currentValue,
+                    new ValueTransformer.TransformerContext<>(holder, field.type, currentValue, field.fieldName,
+                            writeClientFields, fullSync));
+
+        } catch (Exception e) {
+            GTCEu.LOGGER.error("Sync: Failed to serialize field {}", field.fieldName, e);
+        }
+
+        return new CompoundTag();
+    }
+
+    @SuppressWarnings("unchecked")
+    private static void deserializeField(Object holder, FieldSyncData field,
+                                        @Nullable Tag newValue,
+                                        boolean readingClientFields) {
+        if (newValue == null || newValue instanceof CompoundTag compound && compound.isEmpty()) return;
+
+        if (newValue instanceof CompoundTag compound && compound.getBoolean("null")) {
+            field.handle.set(holder, null);
+            return;
+        }
+
+        if (field.transformer == null) {
+            field.setTransformer(ValueTransformers.get(field.type.getRawType()));
+            if (field.transformer == null) {
+                GTCEu.LOGGER.error("Sync: Failed to deserialize field {} in class {}: Missing value transformer for {}",
+                        field.fieldName, holder.getClass().getName(), field.type);
+                return;
+            }
+        }
+
+        try {
+            ValueTransformer<Object> transformer = (ValueTransformer<Object>) field.transformer;
+            var current = field.handle.get(holder);
+
+            Object result = transformer.deserializeNBT(newValue, new ValueTransformer.TransformerContext<>(
+                    holder, field.type, current, field.fieldName, readingClientFields, false));
+
+            if (result != current) {
+                field.handle.set(holder, result);
+            }
+
+        } catch (Exception e) {
+            if (e instanceof UnsupportedOperationException) {
+                GTCEu.LOGGER.error(
+                        "Sync: failed to perform VarHandle set: unsupported op on {} (you are probably trying to sync a final field)",
+                        field.fieldName);
+                return;
+            }
+            GTCEu.LOGGER.error("Sync: Failed to deserialize field {}", field.fieldName, e);
+        }
+    }
+
     public static class SyncManagedTransformer implements ValueTransformer<ISyncManaged> {
 
         @Override
@@ -109,6 +184,12 @@ public class SyncDataHolder {
         @Override
         public @Nullable ISyncManaged deserializeNBT(Tag tag, TransformerContext<ISyncManaged> context) {
             ISyncManaged syncManaged = context.currentValue();
+            var clazz = context.type().getClassValue();
+
+            if (syncManaged == null && clazz != null && ISyncManaged.class.isAssignableFrom(clazz)) {
+                var ctor = ClassSyncData.getClassData(clazz).getClientsideConstructor();
+                if (ctor != null) syncManaged = (ISyncManaged)ctor.get();
+            }
 
             if (syncManaged == null) {
                 GTCEu.LOGGER.error("Sync: ISyncManaged field was null, cannot instantiate {}",
