+++ src/main/java/com/gregtechceu/gtceu/api/sync_system/managed/ISyncManaged.java
@@ -9,20 +9,16 @@ import org.jetbrains.annotations.Nullable;
 
 /**
  * Represents a class with fields that have sync annotations. <br>
- * Differs from {@link ISyncAnnotated} in that more control is provided over syncing. <br>
- * An {@link ISyncManaged} class manages the sync status of itself and its fields,
- * while a {@link ISyncAnnotated} must be managed by a field in an {@link ISyncManaged} class.
  * <p>
  * A field of type {@code T} can be marked with sync annotations if:
  * <ul>
  * <li>{@code T} is primitive
  * <li>{@code T} has an {@link ValueTransformer} registered
  * <li>{@code T} implements {@link INBTSerializable}
- * <li>{@code T} is an {@link ISyncManaged} or {@link ISyncAnnotated} class
+ * <li>{@code T} is an {@link ISyncManaged} class
  * </ul>
  *
  * @see SyncDataHolder
- * @see ISyncAnnotated
  */
 public interface ISyncManaged {
